ENV NOW LOADED BEFORE USE
I noticed that I was trying to set the GRADIO_PORT before I
had actually loaded the environment. The previous code works
in a docker context because the environment is already set
prior to this, but if running this via poetry outside of
docker the GRADIO_PORT environment variable wasn't being
resolved correctly.