Add docker support (#1)
* ADDED DOTENV, MADE GRADIO_PORT VARIABLE, MOVED DEMO INSTANCE INTO CONTEXT MANAGER SCOPE

added support for .env files

added support to make server_port variable at runtime. defaults to the gradio default of 7860

When trying to add support for .env files, and making the server_port variable, I noticed that
 the demo context manager instance was being used outside its scope.

* ADDED SUPPORT FOR DOCKER AND UPDATED README

Added support for running this with Docker and also updated the README.md

* ENV NOW LOADED BEFORE USE

I noticed that I was trying to set the GRADIO_PORT before I
had actually loaded the environment. The previous code works
in a docker context because the environment is already set
prior to this, but if running this via poetry outside of
docker the GRADIO_PORT environment variable wasn't being
resolved correctly.